[finishes #172042836] Assign ticket to admin
