+++ src/routes/ticket.ts
@@ -5,5 +5,6 @@ import Authenticate from "../middleware/authenticate";
 const router = Router();
 
 router.post("/", Authenticate.authenticateUser, TicketController.newTicket);
+router.post("/:id/assign", Authenticate.authenticateUser, TicketController.assignTicket);
 
 export default router;
