+++ src/tests/testMock.ts
@@ -1,8 +1,15 @@
 import jwt from 'jsonwebtoken';
+import { config } from 'dotenv';
+
+config()
 
 const { SECRET } = process.env;
 
 export const userToken = jwt.sign({ userId: 3, username: 'elNinoperezop', role: 'user' }, SECRET, {
+  expiresIn: '1d',
+});
+
+export const secondUserToken = jwt.sign({ userId: 2, username: 'elNino', role: 'user' }, SECRET, {
   expiresIn: '1d',
 });
