+++ src/utils/helper.ts
@@ -1,4 +1,5 @@
 export const validRoles = ['admin', 'user'];
 export const validTypes = ['bug', 'chore', 'feature'];
 export const validAssignee = ['adminOne', 'adminTwo', 'unassigned'];
+export const validAdmins = ['adminOne', 'adminTwo'];
 export const validComplexity = [0, 1, 2, 3];
